feature: multiply - add initial implementation to handle POST /multiply - add necessary dtos - modify http query test file - rename add.HandleAdd to add.Handle
