+++ handlers/multiply/dtos.go
@@ -1,0 +1,6 @@
+package multiply
+
+type MultiplyDto struct {
+	Number1 int `json:"number1"`
+	Number2 int `json:"number2"`
+}
