+++ main.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/ivivanov18/go-cal-api/handlers"
 	"github.com/ivivanov18/go-cal-api/handlers/add"
+	"github.com/ivivanov18/go-cal-api/handlers/multiply"
 )
 
 func main() {
@@ -14,7 +15,8 @@ func main() {
 	router := http.NewServeMux()
 
 	router.HandleFunc("GET /test", handlers.HandleTest)
-	router.HandleFunc("POST /add", add.HandleAdd)
+	router.HandleFunc("POST /add", add.Handle)
+	router.HandleFunc("POST /multiply", multiply.Handle)
 
 	srv := &http.Server{
 		Addr:    ":8080",
